Discard DisposalRequest and add all the types in AddDisposalRequest
